Add files to build a deb src package for ppa upload
This commit adds the files needed to generate a debina source package
suitable for uploading on launchpad to make a PPA (personal package
archive) for Ubuntu.

Due to the way Electron apps are genearlly packaged this is NOT
currently suitable for submission to the actual Debain archives
although it is fine for Launchpad (and I am still exploring ways to
improve it).

An corrected license file will be needed when that is decided upon.

User install instructions need to be added to the main README.md once
the final ppa path is decided on.