Quick fix issue when mangling objects
When the production build is made, js var names are mangled. In
particular, BN class is also mangled but used inside web3 utils
`fromWei` function to validate it can be converted. As the BN class name
is mangled, the util cannot identify the object and breaks. If a string
is passed instead, there are no issues.