Update UI tests
Disclaimer: Almost half the tests were skipped as the whole strategy of UI testing should be re-thought now that the shared logic is in a separate module. Also, newer react-script versions updated Jest and current react-testing-library has evolved a lot form the early version used in this project, breaking the API multiple times, so updating this test would require a bigger effort.